delete status of psy

+++ PSY-master/app/Http/Controllers/AdminController.php
@@ -5,6 +5,7 @@ use Log;
 use App\Models\Years;
 use App\Models\Training_kinds;
 use App\Models\ProfessionalStatus;
+use App\Models\Psychologist;
 
 class AdminController extends Controller {
 protected $request;
@@ -14,7 +15,7 @@ protected $request;
        $this->request = $request;
    }
        
-        public function create() {
+        public function create($error=null) {
             $years  = new Years;
             $is_new = true;
             $years= $years->get();
@@ -25,7 +26,7 @@ protected $request;
             $training_kinds= new Training_kinds();
             $training_kinds= $training_kinds->get();
             $new_training_kind= new Training_kinds();
-            return view( 'forms.admin', compact( 'years', 'is_new', 'new_year','ProfessionalStatus','new_ProfessionalStatus','training_kinds','new_training_kind') );
+            return view( 'forms.admin', compact( 'years', 'is_new', 'new_year','ProfessionalStatus','new_ProfessionalStatus','training_kinds','new_training_kind','error') );
 	}
         
   
@@ -50,22 +51,30 @@ protected $request;
     
         public function destroy( $data) {
             $user_data = \Request::all();
+            $error='';
             if (isset($user_data) && $user_data['type']=='year'){
                 $years= new Years;
                 $years=$years->where('id', '=',$data)->first();
                 $years->delete();
             }
             else if (isset($user_data) && $user_data['type']=='status'){
-                $ProfessionalStatus= new ProfessionalStatus;
-                $ProfessionalStatus=$ProfessionalStatus->where('id', '=',$data)->first();
-                $ProfessionalStatus->delete();
+                $psychologist= new Psychologist();
+                $psychologist=$psychologist->where('professional_status_id', '=',$data)->count();
+                if(isset($psychologist) && $psychologist>0){
+                    $error="שגיאה- לא ניתן למחוק סטאטוס זה מכיוון שקיימות רשומות לפסיכולוגים המכילות סטאטוס זה";
+                }
+                else {
+                    $ProfessionalStatus= new ProfessionalStatus;
+                    $ProfessionalStatus=$ProfessionalStatus->where('id', '=',$data)->first();
+                    $ProfessionalStatus->delete();
+                }
             }
             else{
                  $training_kinds= new Training_kinds();
                  $training_kinds=$training_kinds->where('id', '=',$data)->first();
                  $training_kinds->delete();
             }
-            return $this->create();
+            return $this->create($error);
 	}
